Exported Types

ConvexSampling.SamplingType — Type

SamplingType

Specify the sampling strategy and the number of evaluations of f. Of the type enum.

Possible Inputs

SAMPLE_COMPASS_STAR: (default), uses (2n+1) function evaluations in a compass-star stencil, where n is domain dimension of f.

SAMPLE_SIMPLEX_STAR: uses (n+2) evaluations; this is experimental, and does not currently utilize lambda or epsilon.

Exported Functions

ConvexSampling.eval_sampling_underestimator_coeffs — Function

eval_sampling_underestimator_coeffs(f::Function, xL::T, xU::T;
    samplingPolicy::SamplingType = SAMPLE_COMPASS_STAR, alpha::T,
    lambda::T, epsilon::Float64 = 0.0)

where T is either Vector{Float64} or Float64

Evaluate the coefficients c, b, w0 for which the affine function fAffine(x) = c + dot(b, x - w0) is guaranteed to underestimate f on [xL, xU].

Arguments

f:Function: must be convex and of the form f(x::Vector{Float64})::Float64

xL::T: coordinates for lower bound of box domain on which f is defined

xU::T: coordinates for upper bound of box domain on which f is defined

Keywords

TODO: need to find a way to include all key argument detail

samplingPolicy::SamplingType: an enum specifying sampling strategy. See SamplingType for more details.

lambda::T: an offset of location of domain midpoint, w0. All components must be between (-1.0, 1.0).

alpha::T: dimensionless step length of each sampled point from stencil centre w0.

epsilon::Float64: an absolute error bound for evaluations of f.

Notes

Additional output sR is only used by experimental method SAMPLE_SIMPLEX_STAR.

Examples

To construct the underestimator function as its constant coefficients for function f on box domain xL[i] <= x[i] <= xU[i]:

w0, b, c = eval_sampling_underestimator_coeffs(f, xL, xU)

in which case fAffine(x) == c + dot(b, x - w0) for all x inputs.

ConvexSampling.construct_sampling_underestimator — Function

construct_sampling_underestimator(f::Function, xL::T, xU::T;
    samplingPolicy::SamplingType = SAMPLE_COMPASS_STAR, alpha::T,
    lambda::T, epsilon::Float64 = 0.0)

where T is either Vector{Float64} or Float64

Return affine underestimator function of the format fAffine(x) = c + dot(b, x - w0) by sampling function f at 2n+1 domain points where n is the function dimension.

See eval_sampling_underestimator_coeffs for more details on function inputs.

Example

To construct the underestimator function for the function f on box domain xL[i] <= x[i] <= xU[i] for all x inputs:

julia> 2+2
5

fAffine(x) = constructsamplingunderestimator(f, xL, xU)

ConvexSampling.eval_sampling_underestimator — Function

eval_sampling_underestimator(f::Function, xL::T, xU::T, xIn::T;
samplingPolicy::SamplingType = SAMPLE_COMPASS_STAR, alpha::T,
lambda::T, epsilon::Float64 = 0.0)

where T is either Vector{Float64} or Float64

Evaluate underestimator fAffine constructed by construct_sampling_underestimator at a domain point xIn. That is, yOut = fAffine(xIn).

See eval_sampling_underestimator_coeffs for more details on function inputs.

ConvexSampling.eval_sampling_lower_bound — Function

eval_sampling_lower_bound(f::Function, xL::T, xU::T;
samplingPolicy::SamplingType = SAMPLE_COMPASS_STAR, alpha::T,
lambda::T, epsilon::Float64 = 0.0)

where T is either Vector{Float64} or Float64

Compute the scalar lower bound of f on the interval [xL, xU], so that f(x) >= fL for each x in the box.

See eval_sampling_underestimator_coeffs for more details on function inputs.

Example

fL = eval_sampling_lower_bound(f, xL, xU)

ConvexSampling.plot_sampling_underestimator — Function

plot_sampling_underestimator(f::Function, xL::T, xU::T;
samplingPolicy::SamplingType = SAMPLE_COMPASS_STAR, alpha::T,
lambda::T, epsilon::Float64 = 0.0,
plot3DStyle::Vector = [surface!, wireframe!, surface],
fEvalResolution::Int64 = 10)

where T is either Vector{Float64} or Float64

Plot (1) function, f, (2) affine underestimator, fAffine, and (3) lower bound fL on the box domain [xL, xU].

Additional Keywords

plot3DStyle::Vector: sets the plot style (ex. wireframe, surface, etc.) of each individual plot component in the set order: (1) lower bound, (2) affine under-estimator, (3) convex function.

fEvalResolution::Int64: number of mesh rows per domain dimension in the resulting plot.

Notes

f must be a function of either 1 or 2 variables and must take a Vector{Float64} input. The produced graph may be stored to a variable and later retrieved with @show.

See eval_sampling_underestimator_coeffs for more details on function inputs.